# pkg/pubsub: Don't use time.After if there is no timeout