Here is benchmark comparison in case when deny condition is only buffer size of channel(timeout disabled):
```
benchmark           old ns/op      new ns/op      delta
BenchmarkPubSub     1045873090     1003671134     -4.04%

benchmark           old allocs     new allocs     delta
BenchmarkPubSub     151053         1053           -99.30%

benchmark           old bytes     new bytes     delta
BenchmarkPubSub     9619360       19360         -99.80%
```
Also I added docstring to main datastructure of package.